requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
filelock==3.12.2
idna==3.4
Jinja2==3.1.2
joblib==1.3.1
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.25.2
opencv-python==4.8.0.74
Pillow==10.0.0
requests==2.31.0
scikit-learn==1.3.0
scipy==1.11.1
sympy==1.12
threadpoolctl==3.2.0
torch==2.0.1
torchvision==0.15.2
typing_extensions==4.7.1
urllib3==2.0.4
